Less autoboxing

+++ aws/src/test/java/org/trellisldp/ext/aws/S3BinaryServiceTest.java
@@ -73,7 +73,7 @@ public class S3BinaryServiceTest {
         assertDoesNotThrow(svc.setContent(BinaryMetadata.builder(identifier).mimeType("text/plain").build(),
                     input, emptyMap())::join);
         svc.get(identifier).thenAccept(binary -> {
-            assertEquals((Long) 22L, binary.getSize());
+            assertEquals(22L, binary.getSize());
             try {
                 assertEquals("A sample binary file.", IOUtils.toString(binary.getContent(), UTF_8).trim());
             } catch (final IOException ex) {
@@ -82,7 +82,7 @@ public class S3BinaryServiceTest {
         }).join();
 
         svc.get(identifier).thenAccept(binary -> {
-            assertEquals((Long) 22L, binary.getSize());
+            assertEquals(22L, binary.getSize());
             try {
                 assertEquals("ple bi", IOUtils.toString(binary.getContent(5, 10), UTF_8).trim());
             } catch (final IOException ex) {
